+++ src/control.rs
@@ -1,19 +1,8 @@
 //! Various functions related to the RTC control that are not specifically 
 //! datetime-, timer-, alarm- or clock output-related will be defined here
 
-use super::{PCF8563, DEVICE_ADDRESS, hal, Error, Register, BitFlags, Control, encode_bcd, decode_bcd};
+use super::{PCF8563, DEVICE_ADDRESS, hal, Error, Register, BitFlags, Control, TimerFreq, encode_bcd, decode_bcd};
 use hal::blocking::i2c::{Write, WriteRead};
-
-/// The two possible output types
-#[allow(non_camel_case_types)]
-#[derive(Copy, Clone, Debug)]
-pub enum InterruptOutput {    
-    /// Active when TF active
-    Continuous, 
-    /// Pulsating according to the datasheet
-    Pulsating,     
-}
-
 
 impl<I2C, E> PCF8563<I2C>
 where
@@ -22,31 +11,22 @@ where
     
     /// Enable or disable external clock test mode
     pub fn control_ext_clk_test_mode(&mut self, flag: Control) -> Result<(), Error<E>> {
-        match flag {
-            /// Enable external clock test mode
+        match flag {            
             Control::On => {
                 self.set_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TEST1)       
-            }
-            /// Disable external clock test mode (normal operation)
+            }            
             Control::Off => {
                 self.clear_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TEST1)
             }
         }
     }
 
-    /*
-    /// Enable external clock test mode
-   pub fn enable_ext_clk_test_mode(&mut self) -> Result<(), Error<E>> {
-       self.set_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TEST1)
+    /// Is external clock test mode enabled?
+    pub fn is_ext_clk_mode_enabled(&mut self) -> Result<bool, Error<E>> {
+        self.is_register_bit_flag_high(Register::CTRL_STATUS_1, BitFlags::TEST1)
     }
 
-    /// Disable external clock test mode (normal operation)
-   pub fn disable_ext_clk_test_mode(&mut self) -> Result<(), Error<E>> {
-       self.clear_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TEST1)
-    }
-    */
-
-    /// Start/stop the clock 
+    /// Start/stop the internal clock
     pub fn control_clock(&mut self, flag: Control) -> Result<(), Error<E>> {
         match flag {
             Control::On => {
@@ -58,68 +38,51 @@ where
         }
     }
 
-    /*
-    /// RTC source clock runs
-   pub fn start_clock(&mut self) -> Result<(), Error<E>> {
-       self.clear_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::STOP)
+    /// Is the internal clock running? (0 - running, 1 - stopped)
+    pub fn is_clock_running(&mut self) -> Result<bool, Error<E>> {
+        let flag = self.is_register_bit_flag_high(Register::CTRL_STATUS_1, BitFlags::STOP)?;
+        let flag = flag^true;
+        Ok(flag)
     }
-
-    /// RTC source clock stops
-   pub fn stop_clock(&mut self) -> Result<(), Error<E>> {
-       self.set_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::STOP)
-    }
-    */
 
     /// Enable or disable power-on-reset override facility
      pub fn control_power_on_reset_override(&mut self, flag: Control) -> Result<(), Error<E>> {
-        match flag {
-            /// Enable power-on-reset override
+        match flag {            
             Control::On => {
                 self.set_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TESTC)       
-            }
-            /// Disable power-on-reset override (normal operation)
+            }            
             Control::Off => {
                 self.clear_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TESTC)
             }
         }
     }
 
-    /*
-    /// Enable power-on-reset override facility
-   pub fn enable_power_on_reset_override(&mut self) -> Result<(), Error<E>> {
-       self.set_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TESTC)
+    /// Is power-on-reset override facility enabled?
+    pub fn is_power_on_reset_override_enabled(&mut self) -> Result<bool, Error<E>> {
+        self.is_register_bit_flag_high(Register::CTRL_STATUS_1, BitFlags::TESTC)
     }
 
-    /// Disable power-on-reset override (normal operation)
-   pub fn disable_power_on_reset_override(&mut self) -> Result<(), Error<E>> {
-       self.clear_register_bit_flag(Register::CTRL_STATUS_1, BitFlags::TESTC)
-    }
-
-    */
-
-   /// Interrupt output when TF is active (continuous or pulsating)
-   pub fn timer_interrupt_output(&mut self, output: InterruptOutput) -> Result<(), Error<E>> {
-       match output {
-           InterruptOutput::Continuous => {
-               self.clear_register_bit_flag(Register::CTRL_STATUS_2, BitFlags::TI_TP)       
-           }
-           InterruptOutput::Pulsating => {
-               self.set_register_bit_flag(Register::CTRL_STATUS_2, BitFlags::TI_TP)
-           }
-       }
-   }
-
-   /// Check the status of the Voltage Low detector flag
-   pub fn get_voltage_low_flag(&mut self) -> Result<bool, Error<E>> {
+    /// Check the status of the Voltage Low detector flag
+    pub fn get_voltage_low_flag(&mut self) -> Result<bool, Error<E>> {
        self.is_register_bit_flag_high(Register::VL_SECONDS, BitFlags::VL)
-   }
+    }
    
-   /// Clear voltage low detector flag
-   pub fn clear_voltage_low_flag(&mut self) -> Result<(), Error<E>> {
+    /// Clear voltage low detector flag
+    pub fn clear_voltage_low_flag(&mut self) -> Result<(), Error<E>> {
        self.clear_register_bit_flag(Register::VL_SECONDS, BitFlags::VL)
-   }
+    }
 
-   // pub fn rtc_init()
+    /// Initialize the RTC by setting all the control flags to zero,
+    /// Disabling alarms and timer
+    /// Timer clock set to the lowest frequency for power saving
+    pub fn rtc_init(&mut self) -> Result<(), Error<E>> {
+        self.write_register(Register::CTRL_STATUS_1, 0)?; // clear all the control bits 
+        self.write_register(Register::CTRL_STATUS_2, 0)?;
+        self.clear_voltage_low_flag()?; // clear the low voltage flag
+        self.disable_all_alarms()?; // disable alarm for all the components
+        self.set_timer_frequency(TimerFreq::Timer_1_60Hz)?; // set timer frequency to 1/60 Hz 
+        Ok(())
+    }
 
 }