+++ CHANGELOG.md
@@ -9,15 +9,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ...
 
-## [0.0.3] - 2021-04-04
+## [0.0.4] - 2021-04-04
 
 ### Added
-- Function for setting time only, with the date left unchanged
 - Get and clear Voltage Low detector flag
+- Get external clock mode flag status
+- Get main clock status (running/stopped)
+- Get power-on-reset-override flag status
+- `rtc_init` function (clears various control bits, sets timer frequency to power saving mode)
+- `disable_all_alarms` wrapped to shut off all alarms at once
+- Get timer flag status
+- Get timer interrupt status
+- Get alarm flag status
+- Get alarm interrupt status
+- Read the alarm minutes, hours, day and weekday 
+- Read the current timer value
+- Read century value (0 or 1)
+- Set century (0 or 1)
+- Get clock output status (enabled/disabled)
 
 ### Changed
 - Control enum fields changed to On/Off
-- control_clock() function also uses On/Off instead of Start/Stop
+- `control_clock()` function also uses On/Off instead of Start/Stop
+- Century is set to zero when month is set
+  
+
+## [0.0.3] - 2021-04-03
+
+### Added
+- Function for setting time only, with the date left unchanged
 
 ## [0.0.2] - 2021-04-03
 
@@ -40,6 +60,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Enabling/disabling clock output
 - Setting clock output frequency
 
+[0.0.4]: https://github.com/nebelgrau77/pcf8563-rs/releases/tag/v0.0.4
 [0.0.3]: https://github.com/nebelgrau77/pcf8563-rs/releases/tag/v0.0.3
 [0.0.2]: https://github.com/nebelgrau77/pcf8563-rs/releases/tag/v0.0.2
 [0.0.1]: https://github.com/nebelgrau77/pcf8563-rs/releases/tag/v0.0.1